refactor: split concentration response metrics
